rename graphdb to callable indexer
