+++ analyzer/callable-indexer/pom.xml
@@ -63,10 +63,10 @@
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                                     <manifestEntries>
-                                        <Main-Class>eu.fasten.analyzer.graphplugin.Main</Main-Class>
+                                        <Main-Class>eu.fasten.analyzer.callableindexer.Main</Main-Class>
                                         <X-Compile-Source-JDK>11</X-Compile-Source-JDK>
                                         <X-Compile-Target-JDK>11</X-Compile-Target-JDK>
-                                        <Plugin-Class>eu.fasten.analyzer.graphplugin.GraphDatabasePlugin</Plugin-Class>
+                                        <Plugin-Class>eu.fasten.analyzer.callableindexer.CallableIndexer</Plugin-Class>
                                         <Plugin-Id>graph-plugin</Plugin-Id>
                                         <Plugin-Version>0.0.1</Plugin-Version>
                                         <Plugin-Description>Graph Plugin</Plugin-Description>
